requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
PyJWT==2.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
email-validator==1.1.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
WTForms==2.3.3

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
coverage==7.3.2

gspread==5.12.0
google-auth==2.22.0

gunicorn==20.1.0
